+++ src/ledger/KeyValueHelperImpl.cpp
@@ -72,7 +72,8 @@ KeyValueHelperImpl::storeDelete(LedgerKey const& key)
 bool
 KeyValueHelperImpl::exists(LedgerKey const& key)
 {
-    if (cachedEntryExists(key))
+    // if (cached entry == nullptr) entry does not exists
+    if (cachedEntryExists(key) && getCachedEntry(key))
     {
         return true;
     }