+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 
 * script for xdr generation
 
+### Fixed
+
+* Managing fee for not existing account (horizon panics)
+* get exists Entry from cache
+
 ## 3.0.0-x.10
 ## Fixed
 * Fixed unexpected error (`Unexpected state: tring to unlock more then we have in current cap in base asset`) on cancel of sale participation for fixed price sale